+++ src/Closing/ClosedEntriesStoreInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SixtyEightPublishers\AmpClient\Closing;
+
+interface ClosedEntriesStoreInterface
+{
+    public function isClosed(EntryKey $key): bool;
+}