+++ tests/Closing/ClosingManagerTest.php
@@ -1,0 +1,51 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SixtyEightPublishers\AmpClient\Tests\Closing;
+
+use Hamcrest\Matchers;
+use Mockery;
+use SixtyEightPublishers\AmpClient\Closing\ClosedEntriesStoreInterface;
+use SixtyEightPublishers\AmpClient\Closing\ClosingManager;
+use SixtyEightPublishers\AmpClient\Closing\EntryKey;
+use Tester\Assert;
+use Tester\TestCase;
+
+require __DIR__ . '/../bootstrap.php';
+
+final class ClosingManagerTest extends TestCase
+{
+    public function testStoreShouldBeCalledOnIsBannerClosed(): void
+    {
+        $store = Mockery::mock(ClosedEntriesStoreInterface::class);
+        $manager = new ClosingManager($store);
+
+        $store->shouldReceive('isClosed')
+            ->once()
+            ->with(Matchers::equalTo(EntryKey::banner('foo', '1')))
+            ->andReturn(true);
+
+        Assert::true($manager->isBannerClosed('foo', '1'));
+    }
+
+    public function testStoreShouldBeCalledOnIsPositionClosed(): void
+    {
+        $store = Mockery::mock(ClosedEntriesStoreInterface::class);
+        $manager = new ClosingManager($store);
+
+        $store->shouldReceive('isClosed')
+            ->once()
+            ->with(Matchers::equalTo(EntryKey::position('foo')))
+            ->andReturn(true);
+
+        Assert::true($manager->isPositionClosed('foo'));
+    }
+
+    protected function tearDown(): void
+    {
+        Mockery::close();
+    }
+}
+
+(new ClosingManagerTest())->run();